64-bit (x64) migration + NVIDIA DLSS scaffolding for the Vulkan backend
DLSS is x64+Vulkan only, so this first does the long-requested 32->64-bit port,
then adds the DLSS feature on top.

64-bit migration
- VM/syscall ABI widened int->intptr_t across the engine<->module boundary:
  VM_Call/VM_Create/VM_ArgPtr, vm->systemCall/entryPoint, the three engine
  system-call handlers, vmMain/dllEntry/syscall in game+cgame+ui+q3_ui, and the
  VMF float-arg macro (was ((float*)args)[x], which breaks at 8-byte stride).
  The interpreter marshals the 32-bit VM stack into an intptr_t array.
- x64 (Debug|x64, Release|x64) configs added to all 7 projects (quake3, botlib,
  renderer, splines, cgame, game, q3_ui). Excluded from x64: vm_x86.c (x86 JIT;
  vm.c now provides an interpreter-fallback VM_Compile/VM_CallCompiled for !id386),
  ft2/*, bg_lib.c, cg_newDraw.c.
- Real x64 runtime bugs fixed (compiler-driven + boot-traced):
  * Hunk_InitMemory cacheline align cast (int) -> (intptr_t) [corrupted the heap]
  * FS_AddGameDirectory qsort element size 4 -> sizeof(char*) [LLP64 pointer array]
  * Sys_ListFiles/_findfirst handle int -> intptr_t [truncated handle, AV in FS]
  * win_hinstance/win_wndproc passed as "%i"/(int) -> "%p"/void* [garbage WNDPROC
    crashed CreateWindowEx's first callback; the last blocker before the window
    came up under both GL and Vulkan]
  * __asm{int 3} debug-breaks -> __debugbreak(); x86 FPU/CPUID asm in win_shared.c
    guarded behind id386 with portable fallbacks; CPUSTRING for _M_X64;
    SetWindowLong/GWL_WNDPROC -> SetWindowLongPtr/GWLP_WNDPROC.
  The x64 build now initialises FS, the VM (native DLLs), server, client, the
  Vulkan window+instance+renderer and sound; a residual q3_ui DLL-internal bug
  remains during UI_INIT.

NVIDIA DLSS
- r_dlss cvar (0=Off,1=Quality,2=Balanced,3=Performance,4=Ultra Performance) +
  a "NVIDIA DLSS" spin control in the q3_ui video menu (Vulkan-only, mirrors the
  antialiasing control).
- renderer/vk_dlss.{c,h}: the full NGX Vulkan integration (extension query, init,
  optimal-settings, create feature, per-frame evaluate, Halton jitter, shutdown),
  isolated behind USE_DLSS. The SDK is proprietary and not vendored, so by default
  it builds as portable stubs and r_dlss still renders the scene at a sub-display
  resolution (VK_AA_DLSS) and upscales to the swapchain via the existing fullscreen
  pass -- a real speed-up even without an RTX GPU; the NGX neural evaluate is the
  documented upgrade once the SDK + a motion-vector buffer are present.
- renderer/nvsdk_ngx/DLSS_VULKAN_REFERENCE.md documents the exact NGX API and how
  to drop in the SDK to activate the neural path.

+++ code/q3_ui/ui_main.c
@@ -40,7 +40,7 @@ This is the only way control passes into the module.
 This must be the very first function compiled into the .qvm file
 ================
 */
-int vmMain( int command, int arg0, int arg1, int arg2, int arg3, int arg4, int arg5, int arg6, int arg7, int arg8, int arg9, int arg10, int arg11  ) {
+intptr_t vmMain( intptr_t command, intptr_t arg0, intptr_t arg1, intptr_t arg2, intptr_t arg3, intptr_t arg4, intptr_t arg5, intptr_t arg6, intptr_t arg7, intptr_t arg8, intptr_t arg9, intptr_t arg10, intptr_t arg11  ) {
 	switch ( command ) {
 	case UI_GETAPIVERSION:
 		return UI_API_VERSION;
